+++ test/spec/fixtures/privileges.sql
@@ -19,6 +19,7 @@ GRANT ALL ON TABLE
     , items3
     , "articleStars"
     , articles
+    , arrays
     , auto_incrementing_pk
     , clients
     , comments
@@ -27,6 +28,7 @@ GRANT ALL ON TABLE
     , compound_pk_view
     , deferrable_unique_constraint
     , empty_table
+    , fav_numbers
     , has_count_column
     , has_fk
     , insertable_view_with_join